fix persona streaming
